+++ tool/travis.sh
@@ -1,9 +1,6 @@
 #!/bin/bash
 # Created with https://github.com/dart-lang/mono_repo
 
-# Fast fail the script on failures.
-set -e
-
 if [ -z "$PKG" ]; then
   echo -e '\033[31mPKG environment variable must be set!\033[0m'
   exit 1
@@ -15,7 +12,9 @@ if [ "$#" == "0" ]; then
 fi
 
 pushd $PKG
-pub upgrade
+pub upgrade || exit $?
+
+EXIT_CODE=0
 
 while (( "$#" )); do
   TASK=$1
@@ -23,22 +22,24 @@ while (( "$#" )); do
   dartanalyzer) echo
     echo -e '\033[1mTASK: dartanalyzer\033[22m'
     echo -e 'dartanalyzer --fatal-infos --fatal-warnings .'
-    dartanalyzer --fatal-infos --fatal-warnings .
+    dartanalyzer --fatal-infos --fatal-warnings . || EXIT_CODE=$?
     ;;
   dartfmt) echo
     echo -e '\033[1mTASK: dartfmt\033[22m'
     echo -e 'dartfmt -n --set-exit-if-changed .'
-    dartfmt -n --set-exit-if-changed .
+    dartfmt -n --set-exit-if-changed . || EXIT_CODE=$?
     ;;
   test) echo
     echo -e '\033[1mTASK: test\033[22m'
     echo -e 'pub run test --run-skipped'
-    pub run test --run-skipped
+    pub run test --run-skipped || EXIT_CODE=$?
     ;;
   *) echo -e "\033[31mNot expecting TASK '${TASK}'. Error!\033[0m"
-    exit 1
+    EXIT_CODE=1
     ;;
   esac
 
   shift
 done
+
+exit $EXIT_CODE